% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppcheckR.R
\name{shinyCppcheck}
\alias{shinyCppcheck}
\title{Shiny application to check C/C++}
\usage{
shinyCppcheck()
}
\value{
Nothing, this just launches a Shiny app.
}
\description{
Run a shiny application to check C/C++ files.
}
\note{
The packages listed in the \strong{Suggests} field of the
  package description are required.
}
