% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2json.R
\name{xml2json}
\alias{xml2json}
\title{Title}
\usage{
xml2json(
  xml,
  spaces = 2L,
  linebreaks = TRUE,
  replacer = NULL,
  attributeNamePrefix = "@_",
  textNodeName = "#text",
  ignoreAttributes = FALSE,
  ignoreNameSpace = FALSE,
  parseNodeValue = TRUE,
  parseAttributeValue = TRUE,
  trimValues = TRUE
)
}
\arguments{
\item{trimValues}{}
}
\value{

}
\description{
Title
}
\examples{
xml <- system.file("extdata", "order-schema.xml", package = "xml2")
cat(xml2json(xml))
#
js <- c(
  'if(key === "@_type"){',
  '  return undefined;',
  '} else if(key === "@_name"){',
  '  return value.toUpperCase();',
  '}',
  'return value;'
)
cat(xml2json(xml, linebreaks = TRUE, replacer = js))
}
