% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppcheckR.R
\name{cppcheckR-shiny}
\alias{cppcheckR-shiny}
\alias{cppcheckROutput}
\alias{renderCppcheckR}
\title{Shiny bindings for \code{cppcheckR}}
\usage{
cppcheckROutput(outputId, width = "100\%", height = "400px")

renderCppcheckR(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended}

\item{expr}{an expression that generates a '\code{\link{cppcheckR}}' widget}

\item{env}{the environment in which to evaluate \code{expr}}

\item{quoted}{logical, whether \code{expr} is a quoted expression (with
\code{quote()})}
}
\value{
\code{cppcheckROutput} returns an output element that can be
  included in a Shiny UI definition, and \code{renderCppcheckRR} returns a
  \code{shiny.render.function} object that can be included in a Shiny server
  definition.
}
\description{
Output and render functions for using \code{cppcheckR} within
  Shiny applications and interactive Rmd documents.
}
